GameEngine VS 2010 project files - removed for VS 2010 unnecessary dependencies and build events. Instead added missing project references. Added some ignores.
git-svn-id: http://hcm-lab.de/public/Horde3D/trunk@719 5ce291ac-9df0-446f-9e4f-d57731c4dda7

+++ Tools/GameEngine/src/GameCollisionComponent/GameEngine_Collisions.cpp
@@ -50,7 +50,7 @@ namespace GameEngine
 		if( entity && ( component = static_cast<CollisionComponent*>(entity->component("CollisionHandler")) ) != 0 )
 		{
 			size_t size = component->collisionList()->size();
-			if( index < size )
+			if( (unsigned)index < size )
 				return component->collisionList()->at(index);
 		}	
 		return 0;
